Rework product creation (#45)
* Remove product ID from the product schema

* Add function for dumping a product spec as a dictionary

* Implement PredictionProduct schema

* Remove Product.create() function

* Move internal schemas to dtos module and make schemas module public

* Allow specifying products with and without extended metadata

* Update schemas for afp-product-schemas updates

* Add URL verfication

* Update intent hash generation

* Split schemas into aliased and non-aliased types

* Temporarily remove support for full products with extended metadata

* Rename 'parse' to 'validate'

* Add method for retrieving product specification from the clearing system

* Update product creation example

* Update the list of supported Python versions

+++ afp/hashing.py
@@ -28,9 +28,9 @@ def generate_intent_hash(
         "uint256",  # nonce
         "address",  # trading_protocol_id
         "bytes32",  # product_id
-        "uint256",  # limit_price
+        "int256",  # limit_price
         "uint256",  # quantity
-        "uint256",  # max_trading_fee_rate
+        "uint32",  # max_trading_fee_rate
         "uint256",  # good_until_time
         "uint8",  # side
         "address",  # referral
